requirements.txt
amqp==2.4.2
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
Django==1.11
django-cors-headers==3.0.1
django-redis==4.10.0
djangorestframework==3.9.2
kombu==4.5.0
Pillow==6.0.0
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
vine==1.3.0
